Review-service add-analytics: (3) tests are completed

+++ src/main/java/ru/practicum/workshop/reviewservice/dto/analytics/EventIndicators.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 public class EventIndicators {
     protected Long eventId;
     protected Integer numberOfReviews;
-    protected Integer positiveReviewsPercent;
-    protected Integer negativeReviewsPercent;
+    protected Double positiveReviewsPercent;
+    protected Double negativeReviewsPercent;
 }
